% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addmale.R
\name{addmale}
\alias{addmale}
\title{addmale}
\usage{
addmale(datapre, tp, estadios)
}
\arguments{
\item{datapre}{data frame}

\item{tp}{numeric vector of the constant temperature}

\item{estadios}{vector of length variable}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

