% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgeClases.R
\name{AgeClases}
\alias{AgeClases}
\title{AgeClases}
\usage{
AgeClases(cuadro, estadios, hfeno)
}
\arguments{
\item{cuadro}{matrix, predictions of phenology variables by day}

\item{estadios}{character vector, all states}

\item{hfeno}{list, objects about phenology}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

