% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agenorm.R
\name{agenorm}
\alias{agenorm}
\title{agenorm}
\usage{
agenorm(estimor, modelm, xini = 0.1, ni = 20)
}
\arguments{
\item{estimor}{numeric vector, estimated coefficients of the nonlinear model}

\item{modelm}{integer value, non linear model (position)}

\item{xini}{double, initial value of iteration}

\item{ni}{integer value, number of iterations}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

