% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_mort.R
\name{coef_mort}
\alias{coef_mort}
\title{coef_mort}
\usage{
coef_mort(proc, modelm, estimor, stdmortg, modelo, gg, datm, alg, pesos,
  weights, g, modelim)
}
\arguments{
\item{proc}{a character, type of variable (mortal, taza)}

\item{modelm}{a character, non linear model}

\item{estimor}{numeric vector, estimated coefficients of the nonlinear model}

\item{stdmortg}{numeric vector, standar deviation estimated}

\item{modelo}{a character, non linear model}

\item{gg}{a character, equation}

\item{datm}{data frame, temperatures and mortality}

\item{alg}{a character, algorithm}

\item{pesos}{numeric vector}

\item{weights}{numeric vector}

\item{g}{formula, equation}

\item{modelim}{character vector, all stages of life}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

