% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_ovi.R
\name{coef_ovi}
\alias{coef_ovi}
\title{coef_ovi}
\usage{
coef_ovi(modelm, estimor, stdovi, g, frm, ovifemal, alg, pesos, weights)
}
\arguments{
\item{modelm}{integer value, non linear model (position)}

\item{estimor}{numeric vector, estimated coefficients of the nonlinear model}

\item{stdovi}{numeric vector, standar deviation estimated}

\item{g}{formula, equation}

\item{frm}{formula, non linear equation}

\item{ovifemal}{data frame, temperatures and oviposition}

\item{alg}{a character, algorithm}

\item{pesos}{numeric vector}

\item{weights}{numeric vector}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

