% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contar1.R
\name{contar1}
\alias{contar1}
\title{contar1}
\usage{
contar1(vec, est)
}
\arguments{
\item{vec}{numeric vector}

\item{est}{a character}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

