% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conteo.R
\name{conteo}
\alias{conteo}
\title{conteo}
\usage{
conteo(data, est)
}
\arguments{
\item{data}{data frame}

\item{est}{character, current state}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}

