% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dead_func.R
\name{dead_func}
\alias{dead_func}
\title{dead_func}
\usage{
dead_func(proc, modelm, datm, alg, inim, pesos, weights)
}
\arguments{
\item{proc}{a character, type of variable (mortal, taza)}

\item{modelm}{integer value, non linear model (position)}

\item{datm}{data frame, temperatures and mortality}

\item{alg}{a character, algorithm}

\item{inim}{numeric vector, initial values to model}

\item{pesos}{numeric vector}

\item{weights}{numeric vector}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

