% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deter1.R
\name{deter1}
\alias{deter1}
\title{deter1}
\usage{
deter1(cuadro, estadios, hfeno)
}
\arguments{
\item{cuadro}{matrix, predictions of phenology variables by day}

\item{estadios}{character vector, all states}

\item{hfeno}{list, objects about phenology}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

