% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.tim.R
\name{dev.tim}
\alias{dev.tim}
\title{dev.tim}
\usage{
dev.tim(matri, tp)
}
\arguments{
\item{matri}{data frame, counts of insect by days and others variables}

\item{tp}{numeric vector of the constant temperature}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

