% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devath.R
\name{devath}
\alias{devath}
\title{devath}
\usage{
devath(opc, datos, estadios, est, tp, intervalo, modelo, poli = 1)
}
\arguments{
\item{opc}{integer value (1: life table data, 2: Cohort data)}

\item{datos}{data frame}

\item{estadios}{vector of length variable}

\item{est}{a character, name of the state evaluated}

\item{tp}{numeric vector of the constant temperature}

\item{intervalo}{double, scale in days}

\item{modelo}{a character, dicotomic model ('logit','probit', 'cloglog')}

\item{poli}{double, multiplier value of fecundity (only for parasitoids)}
}
\value{
None
}
\description{
Function to obtain a table of frequencies, statistics of the data, and model parameters of presence of change of stage with respect to the temperature and the logarithm of the days, without considering the intercept, and including the adjustment indicators and the intervals confidence.
}
\examples{
datos<-list(1)
estadios1<-list.files(system.file("extdata",package = "ilcym"))[-5]
# order them acording sequence of life cycle
estadios1<-estadios1[c(1,3,5,2,4)]
# separate only the state names
estadios<-gsub(".tsv", "",estadios1)

# upload the cohort data about PTM (potato pest)
for(i in 1:length(estadios1)) # CAMBIO
{
  datos[[i]]<-read.table(system.file("extdata", estadios1[i], package = "ilcym"), header = FALSE)
}

# evaluate an state (egg for example)
est<-estadios[1]

estadiosINMD<-estadios[1:(length(estadios)-2)]
estadiosADLT<-estadios[(length(estadios)-1):length(estadios)]

intervalo<-1
temperatura<-temp(datos,estadios,est)
tp<-temperatura$temperature
opc=1 #cohort data
modelo<-"logit"

distri<-devath(opc,datos,estadios,est,tp,intervalo=intervalo,modelo)
}
\author{
Pablo Carhuapoma Ramos
}
\seealso{
Other example: \code{\link{PlotDevath}}, \code{\link{temp}}
}

