% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distriModelAdults.R
\name{distriModelAdults}
\alias{distriModelAdults}
\title{distriModelAdults}
\usage{
distriModelAdults(vec, k, sizeInmaduros, hfeno)
}
\arguments{
\item{vec}{numeric vector}

\item{k}{integer value, non linear model (position)}

\item{sizeInmaduros}{integer value}

\item{hfeno}{list, objects about phenology}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

