% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrimodeldeim.R
\name{distrimodeldeim}
\alias{distrimodeldeim}
\title{distrimodeldeim}
\usage{
distrimodeldeim(vec, sll, hfeno)
}
\arguments{
\item{vec}{numeric vector}

\item{sll}{integer value}

\item{hfeno}{list, objects about phenology}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

