% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estado.R
\name{Estado}
\alias{Estado}
\title{Estado}
\usage{
Estado(data, estad, est)
}
\arguments{
\item{data}{matriz, observed life table (fluctuating)}

\item{estad}{character vector, all states}

\item{est}{a character}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

