% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findval.R
\name{findval}
\alias{findval}
\title{temp}
\usage{
findval(dat, cate)
}
\arguments{
\item{dat}{data frame.}

\item{cate}{vector of characters.}
}
\value{
character a string; default is 'finished!'.
}
\description{
development time sub function
}
\details{
no details
}
\examples{
# building
}
\author{
Pablo Carhuapoma Ramos
}
\seealso{
Other example: \code{\link{temp}}
}

