% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenActIndex.R
\name{GenActIndex}
\alias{GenActIndex}
\title{GenActIndex}
\usage{
GenActIndex(posic, TempsAll = TempsAll, coords = coords, x1 = x1,
  y1 = y1, modelim = modelim, modelm = modelm, estadios = estadios,
  xi = xi, steps = steps, filtro = NULL, hfeno)
}
\arguments{
\item{posic}{integer value, position of the cell}

\item{TempsAll}{list, objetcs about geographic details of the temperatures}

\item{coords}{matrix, coordinates of the extent (all FLT files have the same extent)}

\item{x1}{numeric vector, range of the longitude position}

\item{y1}{numeric vector, range of the latitude position}

\item{modelim}{integer value, position of the nonlinear model (development rate for inmature states)}

\item{modelm}{integer value, position of the nonlinear model (development rate for adult states)}

\item{estadios}{character vector, all states}

\item{xi}{double, estimated median from female senescence}

\item{steps}{integer value, between 1 and 48}

\item{filtro}{numeric vector, range of the limit temperature}

\item{hfeno}{list, objects about phenology}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

