% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenMatriz.R
\name{GenMatriz}
\alias{GenMatriz}
\title{GenMatriz}
\usage{
GenMatriz(dir1, dir2, ilon, ilat)
}
\arguments{
\item{dir1}{character, path of the minimun temperature in FLT format}

\item{dir2}{character, path of the maximun temperature in FLT format}

\item{ilon}{numeric vector, range of the longitude}

\item{ilat}{numeric vector, range of the latitude}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

