% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafdete.R
\name{grafdete}
\alias{grafdete}
\title{grafdete}
\usage{
grafdete(simudeter, dia, corrx, corry, tam, lgx, lgy)
}
\arguments{
\item{simudeter}{matrix, life parameters estimated by temperature}

\item{dia}{integer value, day evaluated}

\item{corrx}{numeric vector of two values, the axis range for X}

\item{corry}{numeric vector of two values, the axis range for y#' @param riter list, objects about phenology}

\item{tam}{integer value, extent of the plot}

\item{lgx}{double, x position of the legend}

\item{lgy}{double, y position of the legend}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

