% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graffvida.R
\name{graffvida}
\alias{graffvida}
\title{graffvida}
\usage{
graffvida(Tablelife, cuadro, estadios, N, M, hfeno, labx, laby, corrx, corry,
  lgx, lgy, Rs)
}
\arguments{
\item{Tablelife}{matriz, counts for the observed life table}

\item{cuadro}{matrix, predictions of phenology variables by day}

\item{estadios}{character vector, all states}

\item{N}{integer value, number of insects to simulate}

\item{M}{integer value, number of days}

\item{hfeno}{list, objects about phenology}

\item{labx}{a character}

\item{laby}{a character}

\item{corrx}{numeric vector of two values, the axis range for X}

\item{corry}{numeric vector of two values, the axis range for y}

\item{lgx}{double, x position of the legend}

\item{lgy}{double, y position of the legend}

\item{Rs}{numeric vector, sexual ratios by day}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

