% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafmort.R
\name{grafmort}
\alias{grafmort}
\title{grafmort}
\usage{
grafmort(proc, modelm, estimor, g, datm, corrx = NULL, corry = NULL, mini,
  maxi, limit, tam, labx, laby, titulo, grises = FALSE)
}
\arguments{
\item{proc}{a character, type of variable ("mortal", "taza")}

\item{modelm}{a character, non linear model}

\item{estimor}{numeric vector, estimated coefficients of the nonlinear model}

\item{g}{formula, equation}

\item{datm}{data frame, temperatures and mortality}

\item{corrx}{numeric vector of two values, the axis range for X}

\item{corry}{numeric vector of two values, the axis range for Y}

\item{mini}{double, initial value of line range}

\item{maxi}{double, final value of line range}

\item{limit}{a character, plot the confidence interval ("si", "no")}

\item{tam}{integer value, extent of the plot}

\item{labx}{a character}

\item{laby}{a character}

\item{titulo}{a character}

\item{grises}{logical; default is FALSE}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

