% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafovi.R
\name{grafovi}
\alias{grafovi}
\title{grafovi}
\usage{
grafovi(modelm, estimor, g, ovifemal, corrx = NULL, mini, maxi, ldx, ldy,
  sdli, qto, limit, tam, labx = NULL, laby = NULL, titulo = NULL,
  grises = FALSE, xi)
}
\arguments{
\item{modelm}{integer value, non linear model (position)}

\item{estimor}{numeric vector, estimated coefficients of the nonlinear model}

\item{g}{formula, equation}

\item{ovifemal}{data frame, temperatures and oviposition}

\item{corrx}{numeric vector of two values, the axis range for X}

\item{mini}{double, initial value of line range}

\item{maxi}{double, final value of line range}

\item{ldx}{double, x position of the legend}

\item{ldy}{double, y position of the legend}

\item{sdli}{numeric vector, standar deviation estimated}

\item{qto}{double, quantile for a normal distribution (P=0.95, etc)}

\item{limit}{numeric vector, standar deviation estimated}

\item{tam}{integer value, extent of the plot}

\item{labx}{a character}

\item{laby}{a character}

\item{titulo}{a character}

\item{grises}{logical; default is FALSE}

\item{xi}{double, estimated median from female senescence}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

