% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafshape.R
\name{grafshape}
\alias{grafshape}
\title{grafshape}
\usage{
grafshape(model, estshap, datashap, qtt, sdli, corrx = NULL, corry = NULL,
  mini, maxi, coefi, limit, tam, labx = NULL, laby = NULL, titulo = NULL,
  grises = FALSE, scaleY, scaleX, est, estadios)
}
\arguments{
\item{model}{integer value, non linear model (position)}

\item{estshap}{list of some numeric vectors, the parameters values estimated}

\item{datashap}{data frame, estimated rates}

\item{qtt}{double, quantile of a normal distribution (p=0.95)}

\item{sdli}{double, standar deviation estimated}

\item{corrx}{numeric vector of two values, the axis range for X}

\item{corry}{numeric vector of two values, the axis range for y}

\item{mini}{double, minimun value of the temperature}

\item{maxi}{double, maximun value of the temperature}

\item{coefi}{numeric vector, linear regression parameters}

\item{limit}{logical; default is FALSE, when It considers the confidence lines to curve}

\item{tam}{integer value, extent of the plot}

\item{labx}{a character}

\item{laby}{a character}

\item{titulo}{a character}

\item{grises}{logical; default is FALSE}

\item{scaleY}{double, interval for the Y axis range}

\item{scaleX}{double, interval for the X axis range}

\item{est}{a character}

\item{estadios}{character vector, all states}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

