% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafSimDete.R
\name{grafSimDete}
\alias{grafSimDete}
\title{grafSimDete}
\usage{
grafSimDete(matrizOut, estadios, labx, laby, titulo, legx, legy, corrx1, corrx2)
}
\arguments{
\item{matrizOut}{matrix, life parameters estimated by temperature}

\item{estadios}{character vector, all states}

\item{labx}{a character}

\item{laby}{a character}

\item{titulo}{a character}

\item{legx}{double, x position of the legend}

\item{legy}{double, y position of the legend}

\item{corrx1}{double, minimun axis valuefor X}

\item{corrx2}{double, maximun axis valuefor X}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

