% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsimulp.R
\name{graphsimulp}
\alias{graphsimulp}
\title{graphsimulp}
\usage{
graphsimulp(pars, modls = c(1, 1, 1, 1, 1, 1), direc, ecua, nomec, ejex, ejey,
  tecu, grises, tit, ax, ay)
}
\arguments{
\item{pars}{data frame, temperatures and their estimations of life parameters}

\item{modls}{vector of integer values, simple non linear model (positions)}

\item{direc}{a character}

\item{ecua}{logical; default is FALSE}

\item{nomec}{logical; default is FALSE}

\item{ejex}{double, the axis limit for X}

\item{ejey}{double, the axis limit for Y}

\item{tecu}{integer value, size of the equation}

\item{grises}{logical; default is FALSE}

\item{tit}{a character}

\item{ax}{integer value, size of axis font for X}

\item{ay}{integer value, size of axis font for Y}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

