% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.table2.R
\name{life.table2}
\alias{life.table2}
\title{life.table2}
\usage{
life.table2(data, estad, ovi = TRUE)
}
\arguments{
\item{data}{matriz, observed life table (fluctuating)}

\item{estad}{character vector, all states}

\item{ovi}{logical, default is TRUE, if there is a table of oviposition (fluctuating)}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

