% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.tableD.R
\name{life.tableD}
\alias{life.tableD}
\title{life.tableD}
\usage{
life.tableD(cuadro, estad, n, m, hfeno, Rs, Amp = 1)
}
\arguments{
\item{cuadro}{matrix, predictions of phenology variables by day}

\item{estad}{character vector, all states}

\item{n}{integer value, number of insects to simulate}

\item{m}{integer value, number of days}

\item{hfeno}{list, objects about phenology}

\item{Rs}{numeric vector, sexual ratios by day}

\item{Amp}{integer value}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

