% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrizA.R
\name{matrizA}
\alias{matrizA}
\title{matrizA}
\usage{
matrizA(estadios, hfeno, Table, steps, modelim)
}
\arguments{
\item{estadios}{character vector, all states}

\item{hfeno}{list, objects about phenology}

\item{Table}{data frame, temperatures (tmin and tmax)}

\item{steps}{integer value, between 1 and 48}

\item{modelim}{integer value, position of the nonlinear model (development rate)}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

