% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelcomp.R
\name{modelcomp}
\alias{modelcomp}
\title{modelcomp}
\usage{
modelcomp(test, model, matri, parametros, shap.estimados, coefi)
}
\arguments{
\item{test}{a character, dicotomic model selected('logit','probit', 'cloglog')}

\item{model}{integer value, non linear model (position)}

\item{matri}{data frame, counts of insect by days and others variables}

\item{parametros}{numeric vector, the parameters values estimated}

\item{shap.estimados}{data frame}

\item{coefi}{numeric vector, linear regression parameters}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

