% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mort.select.R
\name{mort.select}
\alias{mort.select}
\title{mort.select}
\usage{
mort.select(datos, estadios, num)
}
\arguments{
\item{datos}{data frame}

\item{estadios}{vector of length variable}

\item{num}{integer value}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

