% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.R
\name{mortality}
\alias{mortality}
\title{mortality}
\usage{
mortality(opc, datos, estadios, est, tp)
}
\arguments{
\item{opc}{integer value (1: life table data, 2: Cohort data)}

\item{datos}{data frame}

\item{estadios}{vector of length variable}

\item{est}{a character}

\item{tp}{numeric vector of the constant temperature}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

