% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovi_func.R
\name{ovi_func}
\alias{ovi_func}
\title{ovi_func}
\usage{
ovi_func(modelm, ovifemal, alg, inim, pesos, weights)
}
\arguments{
\item{modelm}{integer value, non linear model (position)}

\item{ovifemal}{data frame, temperatures and oviposition}

\item{alg}{a character, algorithm}

\item{inim}{numeric vector of the initial values for the modeling}

\item{pesos}{numeric vector}

\item{weights}{numeric vector}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

