% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovip.time.R
\name{ovip.time}
\alias{ovip.time}
\title{ovip.time}
\usage{
ovip.time(dat)
}
\arguments{
\item{dat}{data frame}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

