% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oviposicion.R
\name{oviposicion}
\alias{oviposicion}
\title{oviposicion}
\usage{
oviposicion(opc, datos, colum, mediana, estadios, tp = NULL)
}
\arguments{
\item{opc}{integer value (1: life table data, 2: Cohort data)}

\item{datos}{data frame}

\item{colum}{integer value}

\item{mediana}{numeric vector}

\item{estadios}{vector of length variable}

\item{tp}{numeric vector of the constant temperature}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

