% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.area.R
\name{p.area}
\alias{p.area}
\title{p.area}
\usage{
p.area(ilat, R)
}
\arguments{
\item{ilat}{numeric vector, range of the latitude}

\item{R}{double, number of regions to divide}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

