% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARAlife.R
\name{PARAlife}
\alias{PARAlife}
\title{PARAlife}
\usage{
PARAlife(N, estad, ltb, Rs2 = NULL, ovipo = NULL, nfem2 = NULL)
}
\arguments{
\item{N}{integer value, number of simulations}

\item{estad}{character vector, all states}

\item{ltb}{matrix, life table counts}

\item{Rs2}{numeric vector, sexual ratios by day with position}

\item{ovipo}{matrix,table of oviposition (fluctuating)}

\item{nfem2}{double, number of females obtained from the external table (fluctuating)}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

