% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\title{parameters}
\usage{
parameters(N, estad, ltb, poli = 1)
}
\arguments{
\item{N}{integer value, number of simulations}

\item{estad}{character vector, all states}

\item{ltb}{matrix, life table counts}

\item{poli}{double}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

