% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr6mortal.R
\name{pr6mortal}
\alias{pr6mortal}
\title{pr6mortal}
\usage{
pr6mortal(modelm, datm)
}
\arguments{
\item{modelm}{integer value, non linear model (position)}

\item{datm}{data frame, temperatures and mortality}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

