% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prueba.R
\name{prueba}
\alias{prueba}
\title{prueba}
\usage{
prueba(model, datashap, datao, ini, corrx, corry, punt, labx, laby, titulo,
  grises = FALSE)
}
\arguments{
\item{model}{a character, dicotomic model ('logit','probit', 'cloglog')}

\item{datashap}{data frame, estimated rates}

\item{datao}{data frame}

\item{ini}{numeric vector of the initial values for the modeling}

\item{corrx}{numeric vector of two values, the axis range for X}

\item{corry}{numeric vector of two values, the axis range for y}

\item{punt}{integer value, extent of the plot}

\item{labx}{a character}

\item{laby}{a character}

\item{titulo}{a character}

\item{grises}{logical; default is FALSE}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

