% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruebamortal.R
\name{pruebamortal}
\alias{pruebamortal}
\title{pruebamortal}
\usage{
pruebamortal(proc, modelm, datm, inim, corrx, corry, mini, maxi, labx, laby,
  titulo)
}
\arguments{
\item{proc}{a character, type of variable (mortal, taza)}

\item{modelm}{integer value, non linear model (position)}

\item{datm}{data frame, temperatures and mortality}

\item{inim}{numeric vector, initial values to model}

\item{corrx}{numeric vector of two values, the axis range for X}

\item{corry}{numeric vector of two values, the axis range for Y}

\item{mini}{double, initial value of line range}

\item{maxi}{double, final value of line range}

\item{labx}{a character}

\item{laby}{a character}

\item{titulo}{a character}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

