% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruebaovi.R
\name{pruebaovi}
\alias{pruebaovi}
\title{pruebaovi}
\usage{
pruebaovi(modelm, ovifemal, inim, corrx, corry, mini, maxi, labx, laby, titulo)
}
\arguments{
\item{modelm}{integer value, non linear model (position)}

\item{ovifemal}{data frame, temperatures and oviposition}

\item{inim}{numeric vector of the initial values for the modeling}

\item{corrx}{numeric vector of two values, the axis range for X}

\item{corry}{numeric vector of two values, the axis range for Y}

\item{mini}{double, initial value of line range}

\item{maxi}{double, final value of line range}

\item{labx}{a character}

\item{laby}{a character}

\item{titulo}{a character}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

