% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RateI.R
\name{RateI}
\alias{RateI}
\title{RateI}
\usage{
RateI(vec, Table2, Ki, parametrosc, parametrosm = NULL, funciont,
  funcionm = NULL, nmax, steps, J = NA, modelim)
}
\arguments{
\item{vec}{numeric vector}

\item{Table2}{data frame, temperatures (tmin and tmax)}

\item{Ki}{integer value, state position}

\item{parametrosc}{numeric vector, estimated coefficients of the nonlinear model about the development rate}

\item{parametrosm}{numeric vector, estimated coefficients of the nonlinear model about the mortality}

\item{funciont}{formula, development rate equation}

\item{funcionm}{formula, mortality equation}

\item{nmax}{double}

\item{steps}{integer value, between 1 and 48}

\item{J}{integer value, state evaluated (1: , 2:)}

\item{modelim}{integer value, position of the nonlinear model (development rate)}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

