% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RateR.R
\name{RateR}
\alias{RateR}
\title{RateR}
\usage{
RateR(vec, Table2, nmax, steps, ff, paramR)
}
\arguments{
\item{vec}{numeric vector}

\item{Table2}{data frame, temperatures (tmin and tmax)}

\item{nmax}{double}

\item{steps}{integer value, between 1 and 48}

\item{ff}{formula, equation}

\item{paramR}{numeric vector, estimated coefficients of the nonlinear model}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

