% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ratios.R
\name{Ratios}
\alias{Ratios}
\title{Ratios}
\usage{
Ratios(Table, modelim, modelm, estadios, xi, steps)
}
\arguments{
\item{Table}{data frame, temperatures (tmin and tmax)}

\item{modelim}{integer value, position of the nonlinear model (development rate for inmature states)}

\item{modelm}{integer value, position of the nonlinear model (development rate for adult states)}

\item{estadios}{character vector, all states}

\item{xi}{double, estimated median from female senescence}

\item{steps}{integer value, between 1 and 48}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

