% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalc.R
\name{recalc}
\alias{recalc}
\title{recalc}
\usage{
recalc(ini, niv = 1)
}
\arguments{
\item{ini}{numeric vector, initial values to model}

\item{niv}{double, iteration level}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

