% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{shape}
\alias{shape}
\title{shape}
\usage{
shape(model, datashap, datao, ini, coefi)
}
\arguments{
\item{model}{integer value, non linear model (position)}

\item{datashap}{data frame, estimated rates}

\item{datao}{data frame}

\item{ini}{numeric vector of the initial values for the modeling}

\item{coefi}{numeric vector, linear regression parameters}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

