% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulacion.R
\name{simulacion}
\alias{simulacion}
\title{simulacion}
\usage{
simulacion(cuadro, estadios, n, m, hfeno, Rs)
}
\arguments{
\item{cuadro}{matrix, predictions of phenology variables by day}

\item{estadios}{character vector, all states}

\item{n}{integer value, number of insects to simulate}

\item{m}{integer value, number of days}

\item{hfeno}{list, objects about phenology}

\item{Rs}{numeric vector, sexual ratios by day}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

