% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulacionVariosAnios.R
\name{simulacionVariosAnios}
\alias{simulacionVariosAnios}
\title{simulacionVariosAnios}
\usage{
simulacionVariosAnios(matrizOut, vectorN, estadios, estadiosAges)
}
\arguments{
\item{matrizOut}{matrix, life parameters estimated by temperature}

\item{vectorN}{list}

\item{estadios}{character vector, all states}

\item{estadiosAges}{list}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

