% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation2.R
\name{simulation2}
\alias{simulation2}
\title{simulation2}
\usage{
simulation2(Day, estadiosAges, oviFreq, ageClassMatriz, hfeno, cuadro, estadios,
  numIni, sexratio, Rs, steps, severalYear)
}
\arguments{
\item{Day}{integer value, day evaluated}

\item{estadiosAges}{list}

\item{oviFreq}{matrix, predictions of oviposition}

\item{ageClassMatriz}{matrix}

\item{hfeno}{list, objects about phenology}

\item{cuadro}{matrix, predictions of phenology variables by day}

\item{estadios}{character vector, all states}

\item{numIni}{integer value}

\item{sexratio}{double, between 0 to 1, sex ratio is defined in the ilcym project}

\item{Rs}{numeric vector, sexual ratios by day}

\item{steps}{integer value, between 1 and 48}

\item{severalYear}{logical}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

