% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simultemp.R
\name{simultemp}
\alias{simultemp}
\title{simultemp}
\usage{
simultemp(N, sexratio, isFixed, temp, reps, xi, steps, poli, params, ..)
}
\arguments{
\item{N}{integer value, number of simulations}

\item{sexratio}{double, between 0 to 1, sex ratio is defined in the ilcym project}

\item{isFixed}{logical; default is FALSE}

\item{temp}{numeric vector, temperatures}

\item{reps}{vector of integer values, repetitions for each constant temperature}

\item{xi}{double, estimated median from female senescence}

\item{steps}{integer value, between 1 and 48}

\item{poli}{double}

\item{params}{list}

\item{..}{Phenology objetc or arrays}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

