% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistlife0.R
\name{statistlife0}
\alias{statistlife0}
\title{statistlife0}
\usage{
statistlife0(data, estad)
}
\arguments{
\item{data}{matriz, observed life table (fluctuating)}

\item{estad}{character vector, all states}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

