% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taza.R
\name{taza}
\alias{taza}
\title{taza}
\usage{
taza(opc, datos, tp, colum)
}
\arguments{
\item{opc}{integer value (1: life table data, 2: Cohort data)}

\item{datos}{data frame}

\item{tp}{numeric vector of the constant temperature}

\item{colum}{integer value}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

