% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{temp}
\alias{temp}
\title{temp}
\usage{
temp(datos, estadios, est)
}
\arguments{
\item{datos}{data frame.}

\item{estadios}{vector of characters.}

\item{est}{a character.}
}
\value{
character a string; default is 'finished!'.
}
\description{
A special curve using specific parameters and also its modelling
}
\details{
If no estimated values of the parameters of the model must enable "modeling = TRUE", then include the initial values
}
\examples{
# temp description
str(temp)
}
\author{
Pablo Carhuapoma Ramos
}
\seealso{
Other example: \code{\link{PlotDevath}},
  \code{\link{devath}}
}

