% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verif.val.num.R
\name{verif.val.num}
\alias{verif.val.num}
\title{verif.val.num}
\usage{
verif.val.num(datos, cate, estadios)
}
\arguments{
\item{datos}{data frame}

\item{cate}{character}

\item{estadios}{vector of length variable}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

