% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verif.val.R
\name{verif.val}
\alias{verif.val}
\title{verif.val}
\usage{
verif.val(datos, cate)
}
\arguments{
\item{datos}{data frame.}

\item{cate}{a character.}
}
\value{
None.
}
\description{
development time sub function
}
\details{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

