% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIDA.R
\name{VIDA}
\alias{VIDA}
\title{VIDA}
\usage{
VIDA(vidalife, Tablelife, p, cuadro, estadios, N, M, hfeno, Rs, Rs2 = NULL)
}
\arguments{
\item{vidalife}{matriz, observed life table (fluctuating)}

\item{Tablelife}{matriz, counts for the simulated life table}

\item{p}{matrix, life parameters estimated by temperature}

\item{cuadro}{matrix, predictions of phenology variables by day}

\item{estadios}{character vector, all states}

\item{N}{integer value, number of insects to simulate}

\item{M}{integer value, number of days}

\item{hfeno}{list, objects about phenology}

\item{Rs}{numeric vector, sexual ratios by day}

\item{Rs2}{numeric vector, sexual ratios by day with position}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

