% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone.div.R
\name{zone.div}
\alias{zone.div}
\title{zone.div}
\usage{
zone.div(dir1, dir2, ilon, ilat, R, dir.out, name.out = name.out,
  modelim = modelim, modelm = modelm, estadios = estadios, xi = xi,
  steps = steps, filtro = NULL, hfeno)
}
\arguments{
\item{dir1}{character, path of the minimun temperature in FLT format}

\item{dir2}{character, path of the maximun temperature in FLT format}

\item{ilon}{numeric vector, range of the longitude}

\item{ilat}{numeric vector, range of the latitude}

\item{R}{double, number of regions to divide}

\item{dir.out}{character, final path of the folder for the indexes (ASCII files of ERI,GI,AI)}

\item{name.out}{character, name of the ASCII files}

\item{modelim}{integer value, position of the nonlinear model (development rate for inmature states)}

\item{modelm}{integer value, position of the nonlinear model (development rate for adult states)}

\item{estadios}{character vector, all states}

\item{xi}{double, estimated median from female senescence}

\item{steps}{integer value, between 1 and 48}

\item{filtro}{numeric vector, range of the limit temperature}

\item{hfeno}{list, objects about phenology}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

