% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devath.R
\name{devath}
\alias{devath}
\title{devath}
\usage{
devath(opc, datos, estadios, est, tp, intervalo, modelo, poli = 1)
}
\arguments{
\item{opc}{integer value (1: life table data, 2: Cohort data)}

\item{datos}{data frame}

\item{estadios}{vector of length variable}

\item{est}{a character}

\item{tp}{numeric vector of the constant temperature}

\item{intervalo}{double, scale in days}

\item{modelo}{a character, dicotomic model ('logit','probit', 'cloglog')}

\item{poli}{double, multiplier value of fecundity (only for parasitoids)}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

