% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDevath.R
\name{PlotDevath}
\alias{PlotDevath}
\title{PlotDevath}
\usage{
PlotDevath(matri, parametros, test, corrx, lgx, lgy, tam, labx = NULL,
  laby = NULL, titulo = NULL, grises = FALSE, intervalo = intervalo,
  exis.val)
}
\arguments{
\item{matri}{data frame, counts of insect by days and others variables}

\item{parametros}{numeric vector, slopes of modeling outputs}

\item{test}{a character, model selected}

\item{corrx}{numeric vector of two values, the axis range for X}

\item{lgx}{double, x position of the legend}

\item{lgy}{double, y position of the legend}

\item{tam}{integer value, extent of the plot}

\item{labx}{a character}

\item{laby}{a character}

\item{titulo}{a character}

\item{grises}{logical; default is FALSE}

\item{intervalo}{double, scale in days}

\item{exis.val}{numeric vector, if there are live individuals in a temperature}
}
\value{
None
}
\description{
no details
}
\examples{
#building

}
\author{
Pablo Carhuapoma Ramos
}

